Update UI part1
