+++ frontend/package.json
@@ -3,6 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "@fontsource/roboto": "^4.5.8",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.11.16",
@@ -14,6 +16,7 @@
     "axios": "^1.3.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-number-format": "^5.1.4",
     "react-phone-input-2": "^2.13.9",
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
