+++ frontend/src/component/JobDetails.js
@@ -1,0 +1,12 @@
+import { userType } from "../lib/isAuth";
+import { RecruiterJobDetails } from "./recruiter";
+
+export default function JobDetails(props) {
+    if (userType() === "recruiter") {
+        return <RecruiterJobDetails {...props} />;
+    } else if (userType() === "applicant") {
+        return <RecruiterJobDetails {...props} />;
+    } else {
+        return <>Not logged in</>;
+    }
+}
