+++ frontend/src/component/applicant/index.js
@@ -1,0 +1,2 @@
+export { default as ApplicantProfile } from './Profile';
+export { default as ApplicantHome } from './Home';
