+++ frontend/src/component/recruiter/AcceptedApplicants.js
@@ -551,7 +551,7 @@ const ApplicationTile = (props) => {
           </Grid>
           <Grid item>Job Title: {application.job.title}</Grid>
           <Grid item>Role: {application.job.jobType}</Grid>
-          <Grid item>Applied On: {appliedOn.toLocaleDateString()}</Grid>
+          <Grid item>Applied On: {appliedOn.toLocaleDateString('vi-VN')}</Grid>
           <Grid item>
             SOP: {application.sop !== "" ? application.sop : "Not Submitted"}
           </Grid>
