+++ frontend/src/component/recruiter/JobDetails.js
@@ -1,0 +1,273 @@
+import { Grid, Modal, Typography, Paper, Button, TextField } from "@mui/material";
+import { useContext, useEffect, useState } from "react";
+import { useParams } from "react-router-dom";
+import { SetPopupContext } from "../../App";
+import apiList from "../../lib/apiList";
+import axios from "axios";
+import { makeStyles } from "@mui/styles";
+
+const useStyles = makeStyles((theme) => ({
+  body: {
+    height: "inherit",
+  },
+  jobTileOuter: {
+    padding: "30px",
+    margin: "20px 0",
+    boxSizing: "border-box",
+    width: "100%",
+  },
+  popupDialog: {
+    height: "100%",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+}));
+
+export default function JobDetails(props) {
+  const classes = useStyles();
+  const { jobId } = useParams();
+  const [job, setJob] = useState();
+  const setPopup = useContext(SetPopupContext);
+  const [open, setOpen] = useState(false);
+  const [openUpdate, setOpenUpdate] = useState(false);
+  const [jobDetails, setJobDetails] = useState(job);
+
+  const getData = () => {
+    const address = apiList.jobs + "/" + jobId;
+    axios.get(address, {
+      headers: {
+        Authorization: `Bearer ${localStorage.getItem("token")}`,
+      }
+    }).then((res) => {
+      console.log(res.data);
+      setJob(res.data);
+    }).catch((err) => {
+      console.log(err.response.data);
+      setPopup({
+        open: true,
+        severity: "error",
+        message: "Error",
+      });
+    });
+  }
+
+  const handleInput = (key, value) => {
+    setJobDetails({
+      ...jobDetails,
+      [key]: value,
+    });
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  const handleCloseUpdate = () => {
+    setOpenUpdate(false);
+  };
+
+  const handleDelete = () => {
+    console.log(job._id);
+    axios
+      .delete(`${apiList.jobs}/${job._id}`, {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem("token")}`,
+        },
+      })
+      .then((response) => {
+        setPopup({
+          open: true,
+          severity: "success",
+          message: response.data.message,
+        });
+        getData();
+        handleClose();
+      })
+      .catch((err) => {
+        console.log(err.response);
+        setPopup({
+          open: true,
+          severity: "error",
+          message: err.response.data.message,
+        });
+        handleClose();
+      });
+  };
+
+  const handleJobUpdate = () => {
+    axios
+      .put(`${apiList.jobs}/${job._id}`, jobDetails, {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem("token")}`,
+        },
+      })
+      .then((response) => {
+        setPopup({
+          open: true,
+          severity: "success",
+          message: response.data.message,
+        });
+        getData();
+        handleCloseUpdate();
+      })
+      .catch((err) => {
+        console.log(err.response);
+        setPopup({
+          open: true,
+          severity: "error",
+          message: err.response.data.message,
+        });
+        handleCloseUpdate();
+      });
+  };
+
+  const jobModals = () => (
+    <>
+      <Modal open={open} onClose={handleClose} className={classes.popupDialog}>
+        <Paper
+          style={{
+            padding: "20px",
+            outline: "none",
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "center",
+            minWidth: "30%",
+            alignItems: "center",
+          }}
+        >
+          <Typography variant="h4" style={{ marginBottom: "10px" }}>
+            Are you sure?
+          </Typography>
+          <Grid container justify="center" spacing={5}>
+            <Grid item>
+              <Button
+                variant="contained"
+                color="secondary"
+                style={{ padding: "10px 50px" }}
+                onClick={() => handleDelete()}
+              >
+                Delete
+              </Button>
+            </Grid>
+            <Grid item>
+              <Button
+                variant="contained"
+                color="primary"
+                style={{ padding: "10px 50px" }}
+                onClick={() => handleClose()}
+              >
+                Cancel
+              </Button>
+            </Grid>
+          </Grid>
+        </Paper>
+      </Modal>
+      <Modal
+        open={openUpdate}
+        onClose={handleCloseUpdate}
+        className={classes.popupDialog}
+      >
+        <Paper
+          style={{
+            padding: "20px",
+            outline: "none",
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "center",
+            minWidth: "30%",
+            alignItems: "center",
+          }}
+        >
+          <Typography variant="h4" style={{ marginBottom: "10px" }}>
+            Update Details
+          </Typography>
+          <Grid
+            container
+            direction="column"
+            spacing={3}
+            style={{ margin: "10px" }}
+          >
+            <Grid item>
+              <TextField
+                label="Application Deadline"
+                type="datetime-local"
+                value={jobDetails.deadline.substr(0, 16)}
+                onChange={(event) => {
+                  handleInput("deadline", event.target.value);
+                }}
+                InputLabelProps={{
+                  shrink: true,
+                }}
+                variant="outlined"
+                fullWidth
+              />
+            </Grid>
+            <Grid item>
+              <TextField
+                label="Maximum Number Of Applicants"
+                type="number"
+                variant="outlined"
+                value={jobDetails.maxApplicants}
+                onChange={(event) => {
+                  handleInput("maxApplicants", event.target.value);
+                }}
+                InputProps={{ inputProps: { min: 1 } }}
+                fullWidth
+              />
+            </Grid>
+            <Grid item>
+              <TextField
+                label="Positions Available"
+                type="number"
+                variant="outlined"
+                value={jobDetails.maxPositions}
+                onChange={(event) => {
+                  handleInput("maxPositions", event.target.value);
+                }}
+                InputProps={{ inputProps: { min: 1 } }}
+                fullWidth
+              />
+            </Grid>
+          </Grid>
+          <Grid container justify="center" spacing={5}>
+            <Grid item>
+              <Button
+                variant="contained"
+                color="secondary"
+                style={{ padding: "10px 50px" }}
+                onClick={() => handleJobUpdate()}
+              >
+                Update
+              </Button>
+            </Grid>
+            <Grid item>
+              <Button
+                variant="contained"
+                color="primary"
+                style={{ padding: "10px 50px" }}
+                onClick={() => handleCloseUpdate()}
+              >
+                Cancel
+              </Button>
+            </Grid>
+          </Grid>
+        </Paper>
+      </Modal>
+    </>
+  );
+
+  useEffect(getData, []);
+
+  return (
+    <>
+      {job ?
+        <Grid container item direction="column">
+          <Grid item>
+            <Typography variant="h3">{job.title}</Typography>
+          </Grid>
+        </Grid> : null}
+    </>
+
+  );
+}
