+++ frontend/src/component/recruiter/index.js
@@ -1,0 +1,3 @@
+export { default as RecruiterProfile } from './Profile';
+export { default as RecruiterHome } from './Home';
+export { default as RecruiterJobDetails } from './JobDetails';
